pcdaemon/requirements.txt
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.6
MouseInfo==0.1.3
mss==6.1.0
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
PyAudio==0.2.11
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycparser==2.20
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-dotenv==0.19.0
python3-xlib==0.15
PyTweening==1.0.3
six==1.16.0
toml==0.10.2
